InputCoffee added
Split off Input Brew into its own world
Added Input Coffee
