+++ src/components/app-body.js
@@ -2,7 +2,8 @@
 
 import React, {Component} from 'react';
 
-import FancierForm from './input-stuff.js';
+import FancierBrewForm from './brew-input.js';
+import FancierCoffeeForm from './coffee-input.js';
 import AddCoffeeForm from './temporary-add-to-db-component';
 import CoffeeShelf from './test-shelf.js';
 
@@ -44,7 +45,12 @@ class SimpleAppBody extends Component {
 
       <h2 className="card-heading">Add Brew</h2>
       <div className="App-card">
-        <FancierForm />
+        <FancierBrewForm />
+      </div>
+
+      <h2 className="card-heading">Add Coffee</h2>
+      <div className="App-card">
+        <FancierCoffeeForm />
       </div>
 
       <h2 className="card-heading">Add JSON</h2>
